Manejo de errores, Test cobertura 100%., Enum y config

+++ src/main/java/com/indramakers/example/measuresms/services/LocationService.java
@@ -1,5 +1,8 @@
 package com.indramakers.example.measuresms.services;
 
+import com.indramakers.example.measuresms.config.ErrorCodes;
+import com.indramakers.example.measuresms.exceptions.BusinessExceptions;
+import com.indramakers.example.measuresms.exceptions.NotFoundException;
 import com.indramakers.example.measuresms.model.entities.Location;
 import com.indramakers.example.measuresms.model.entities.Measure;
 import com.indramakers.example.measuresms.repositories.IDevicesRepository;
@@ -22,20 +25,27 @@ public class LocationService {
         if(locationRepository.findByLocationName(location.getName()).isEmpty()){
             locationRepository.createLocation(location);
         }else{
-            throw new RuntimeException("Error, el nombre de location indicado ya existe");
+            throw new BusinessExceptions(ErrorCodes.LOCATION_WITH_NAME_EXISTS);
         }
 
     }
 
     public List<Location> getById(int id){
+        if(locationRepository.findByLocationId(id).isEmpty()){
+            throw new NotFoundException(ErrorCodes.NO_LOCATIONS_FOUND.getMessage());
+        }
         return locationRepository.findByLocationId(id);
     }
 
     public void deleteLocation(int id_location){
+        if(locationRepository.findByLocationId(id_location).isEmpty()){
+            throw new NotFoundException(ErrorCodes.NO_LOCATIONS_FOUND.getMessage());
+        }
+
         if (iDevicesRepository.findByIdLocation(id_location).isEmpty()){
             locationRepository.deleteLocation(id_location);
         }else{
-            throw new RuntimeException("Error, hay un device asociado a esa location");
+            throw new BusinessExceptions(ErrorCodes.DEVICE_REFERENCED_TO_LOCATION);
         }
 
     }
@@ -45,6 +55,13 @@ public class LocationService {
     }
 
     public List<Measure> getMeasures(int id_location){
+        if(locationRepository.findByLocationId(id_location).isEmpty()){
+            throw new NotFoundException(ErrorCodes.NO_LOCATIONS_FOUND.getMessage());
+        }
+
+        if(locationRepository.getMeasures(id_location).isEmpty()){
+            throw new NotFoundException(ErrorCodes.NO_MEASURES_FOUND.getMessage());
+        }
         return locationRepository.getMeasures(id_location);
     }
